requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
praw>=7.7.1
requests>=2.31.0
feedparser>=6.0.11
python-dotenv>=1.0.1
